+++ pkgs/universal_storage_sync/example/todo_file_app/lib/state/app_state.dart
@@ -249,6 +249,8 @@ class AppState extends ChangeNotifier {
   Future<void> _initializeStorage() async {
     if (!hasWorkspace) return;
 
+    resolvePlatformDirectoryOfConfig(filePathConfig);
+
     final fileSystemConfig = FileSystemConfig.fromFilePathConfig(
       filePathConfig,
     );