+++ pkgs/universal_storage_sync_utils/lib/universal_storage_sync_utils.dart
@@ -4,8 +4,8 @@ library;
 
 export 'src/folder_picker.dart';
 export 'src/macos_bookmark_manager.dart';
+export 'src/models/pick_result.dart';
 export 'src/path_validator.dart';
 export 'src/repository_manager.dart';
-export 'src/result.dart';
 export 'src/sync_utils/hash_utils.dart';
 export 'src/sync_utils/merge_utils.dart';