+++ pkgs/universal_storage_sync_utils/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-dev.8] - 2025-10-25
+
+- added resolvePlatformDirectoryOfConfig function
+- chore: interface package dependency to 0.1.0-dev.8
+
 ## [0.1.0-dev.7] - 2025-10-25
 
 - replaced dart:io with universal_io